% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-addins.R
\name{insert_examplesTempdir_template}
\alias{insert_examplesTempdir_template}
\title{Insert a temporary working directory template}
\usage{
insert_examplesTempdir_template()
}
\description{
This function inserts a template for the \verb{@examplesTempdir} tag at the cursor position.
It's designed to be used as an RStudio addin.
}
