% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rstudio-addins.R
\name{insert_examplesWebR_template}
\alias{insert_examplesWebR_template}
\title{Insert a webR examples template}
\usage{
insert_examplesWebR_template()
}
\description{
This function inserts a template for the \verb{@examplesWebR} tag at the cursor position.
It's designed to be used as an RStudio addin.
}
