% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocleteer-package.R
\docType{package}
\name{rocleteer-package}
\alias{rocleteer}
\alias{rocleteer-package}
\title{rocleteer: Extra Roclets for roxygen2}
\description{
Provides extensions to the roxygen2 documentation system. The exact extensions are still being added.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://r-pkg.thecoatlessprofessor.com/rocleteer/}
  \item \url{https://github.com/coatless-rpkg/rocleteer}
  \item Report bugs at \url{https://github.com/coatless-rpkg/rocleteer/issues}
}

}
\author{
\strong{Maintainer}: James Joseph Balamuta \email{james.balamuta@gmail.com} (\href{https://orcid.org/0000-0003-2826-8458}{ORCID})

}
\keyword{internal}
