% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examplesWebR.R
\name{webr_repl_href}
\alias{webr_repl_href}
\title{Generate webR REPL link}
\usage{
webr_repl_href(encoded_code, version = "latest")
}
\arguments{
\item{encoded_code}{Base64 encoded R code}

\item{version}{webR version (default: "latest")}
}
\value{
A URL string for the webR REPL
}
\description{
Create a webR REPL link for the given encoded code
}
\keyword{internal}
